requirements.txt
flask
python-dotenv
langchain
langchain-google-genai
